pyproject.toml
[project]
name = "wyn360-cli"
version = "0.4.0"
description = "WYN360 - An intelligent AI coding assistant CLI tool powered by Anthropic Claude"
authors = [
    {name = "Yiqiao Yin", email = "yiqiao.yin@wyn-associates.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "click>=8.1.0",
    "pydantic-ai>=1.13.0",
    "anthropic[bedrock]>=0.39.0",
    "google-genai>=1.0.0",
    "rich>=13.0.0",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "prompt-toolkit>=3.0.0",
    "pyyaml>=6.0.0",
    "huggingface-hub>=0.20.0",
    "crawl4ai>=0.7.6",
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "pytesseract>=0.3.10",
    "pdf2image>=1.16.0",
    "Pillow>=10.0.0",
    "pymupdf-layout>=0.1.0",
    "pymupdf4llm>=0.1.0",
    "cryptography>=42.0.0",
    "playwright>=1.40.0",
    "stagehand-py>=0.1.0",
]

[project.scripts]
wyn360 = "wyn360_cli.cli:main"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
    "mkdocs (>=1.6.1,<2.0.0)",
    "mkdocs-material (>=9.7.0,<10.0.0)",
    "mkdocstrings[python] (>=0.30.1,<0.31.0)"
]
